+++ python/data/format.py
@@ -16,7 +16,7 @@ class Events:
     def __post_init__(self):
         assert self.x.dtype == np.uint16
         assert self.y.dtype == np.uint16
-        assert self.p.dtype == np.uint8
+        assert self.p.dtype == np.int16
         assert self.t.dtype == np.int64
 
         assert self.x.shape == self.y.shape == self.p.shape == self.t.shape
@@ -53,7 +53,7 @@ if __name__ == '__main__':
             Events(
                 x=np.array([0, 1], dtype='uint16'),
                 y=np.array([1, 2], dtype='uint16'),
-                p=np.array([0, 0], dtype='uint8'),
+                p=np.array([0, 0], dtype='int16'),
                 t=np.array([0, 5], dtype='int64')),
             width=2,
             height=3,
